Change long long to int64_t

+++ pyogrio/_ogr.pxd
@@ -220,7 +220,7 @@ cdef extern from "ogr_api.h":
     void            OGR_F_SetFieldDateTime(OGRFeatureH feature, int n, int y, int m, int d, int hh, int mm, int ss, int tz)
     void            OGR_F_SetFieldDouble(OGRFeatureH feature, int n, double value)
     void            OGR_F_SetFieldInteger(OGRFeatureH feature, int n, int value)
-    void            OGR_F_SetFieldInteger64(OGRFeatureH feature, int n, long long value)
+    void            OGR_F_SetFieldInteger64(OGRFeatureH feature, int n, int64_t value)
     void            OGR_F_SetFieldString(OGRFeatureH feature, int n, char *value)
     void            OGR_F_SetFieldBinary(OGRFeatureH feature, int n, int l, unsigned char *value)
     void            OGR_F_SetFieldNull(OGRFeatureH feature, int n)  # new in GDAL 2.2
